Split tests into files depending on form type
